correctif history

+++ FitnessHistory.cpp
@@ -7,7 +7,7 @@ void FitnessHistory::setup(size_t PopulationEngine)
 
 void FitnessHistory::insertNewStatistics(size_t PopulationEngine, FitnessStatistics statistics) {
 
-	int size{ mStatisticsHistory[PopulationEngine].size() };
+	size_t size{ mStatisticsHistory[PopulationEngine].size() };
 	mStatisticsHistory[PopulationEngine].resize(size + 1);
 	mStatisticsHistory[PopulationEngine].at(size) = statistics;
